requirements.txt
anthropic>=0.8.0
matplotlib>=3.7.1
networkx>=3.1
numpy>=1.24.3
openai>=1.12.0
pymilvus>=2.3.6
python-dotenv>=1.0.0
scikit-learn>=1.3.0
streamlit>=1.31.1
voyageai>=0.1.5 
gpt_researcher